+++ apps/web/src/_pages/feed/FeedPage.tsx
@@ -5,7 +5,6 @@ import {
   type KeyboardEvent,
   useDeferredValue,
   useEffect,
-  useMemo,
   useState,
 } from 'react';
 
@@ -30,23 +29,21 @@ import {
 
 import { PostHeader } from '@/features/post';
 
-import { PostBody } from '@/entities/post';
-import { postQueryOptions } from '@/entities/post/model/queries/useGetPosts';
-import type { GetPostsItemResponseDto } from '@/entities/post/model/types';
+import { PostBody, postQueryOptions } from '@/entities/post';
 
 import { QUERY_STALE_TIME } from '@/shared/constants';
 import { useBreakpoint, useFetchNextOnScroll } from '@/shared/hooks';
 import { ActionHeader, IconCountButton } from '@/shared/ui';
 
-type FeedSearchPost = {
-  id: string;
-  author: string;
-  createdAt: string;
-  avatarUrl?: string;
-  content: string;
-  likeCount: number;
-  commentCount: number;
-};
+import {
+  createRecentSearches,
+  type FeedSearchPost,
+  formatEmptyResultMessage,
+  loadRecentSearches,
+  mapPostToCard,
+  saveRecentSearches,
+  updateSearchParam,
+} from '@/_pages/feed/feed-page.helpers';
 
 const COPY = {
   back: '뒤로',
@@ -60,109 +57,6 @@ const COPY = {
   loading: '검색 결과를 불러오는 중...',
   htmlOnlyPost: '이미지가 포함된 게시글입니다.',
 } as const;
-
-const FEED_RECENT_SEARCHES_KEY = 'feed-recent-searches';
-
-const HTML_TAG_PATTERN = /<\/?[a-z][\s\S]*?>/gi;
-const HTML_ENTITY_PATTERN = /&[a-zA-Z0-9#]+;/g;
-
-const decodeHtmlEntity = (value: string) =>
-  value
-    .replaceAll('&nbsp;', ' ')
-    .replaceAll('&amp;', '&')
-    .replaceAll('&lt;', '<')
-    .replaceAll('&gt;', '>')
-    .replaceAll('&quot;', '"')
-    .replaceAll('&#39;', "'");
-
-const getPostPreviewText = (content: string) => {
-  const withoutTags = content.replace(HTML_TAG_PATTERN, ' ');
-  const normalized = withoutTags
-    .replace(HTML_ENTITY_PATTERN, decodeHtmlEntity)
-    .replace(/\s+/g, ' ')
-    .trim();
-
-  return normalized || COPY.htmlOnlyPost;
-};
-
-const saveRecentSearches = (keywords: string[]) => {
-  if (typeof window === 'undefined') return;
-
-  window.localStorage.setItem(
-    FEED_RECENT_SEARCHES_KEY,
-    JSON.stringify(keywords),
-  );
-};
-
-const loadRecentSearches = () => {
-  if (typeof window === 'undefined') return [];
-
-  const saved = window.localStorage.getItem(FEED_RECENT_SEARCHES_KEY);
-
-  if (!saved) return [];
-
-  try {
-    const parsed = JSON.parse(saved);
-
-    if (!Array.isArray(parsed)) return [];
-
-    return parsed.filter(
-      (keyword): keyword is string => typeof keyword === 'string',
-    );
-  } catch {
-    return [];
-  }
-};
-
-const updateSearchParam = ({
-  keyword,
-  pathname,
-  router,
-  searchParams,
-}: {
-  keyword: string;
-  pathname: string;
-  router: ReturnType<typeof useRouter>;
-  searchParams: { toString(): string };
-}) => {
-  const params = new URLSearchParams(searchParams.toString());
-  const trimmedKeyword = keyword.trim();
-
-  if (trimmedKeyword.length === 0) {
-    params.delete('q');
-  } else {
-    params.set('q', trimmedKeyword);
-  }
-
-  const queryString = params.toString();
-  const nextUrl = queryString ? `${pathname}?${queryString}` : pathname;
-
-  router.replace(nextUrl, { scroll: false });
-};
-
-const createRecentSearches = (keyword: string, recentSearches: string[]) => {
-  const trimmedKeyword = keyword.trim();
-
-  if (!trimmedKeyword) return recentSearches;
-
-  return [
-    trimmedKeyword,
-    ...recentSearches.filter((item) => item !== trimmedKeyword),
-  ].slice(0, 8);
-};
-
-const formatEmptyResultMessage = (keyword: string) =>
-  COPY.noSearchResult.replace('%s', keyword);
-
-const mapPostToCard = (post: GetPostsItemResponseDto): FeedSearchPost => ({
-  id: post.postId,
-  author: post.writerNickname,
-  createdAt: post.createdAt,
-  avatarUrl: post.writerProfileImageUrl ?? undefined,
-  content: getPostPreviewText(post.content),
-  likeCount: post.numLike,
-  commentCount: post.numComment,
-});
 
 const FeedSearchInput = ({
   keyword,
@@ -269,7 +163,9 @@ export const FeedPage = () => {
       enabled: hasKeyword,
       staleTime: QUERY_STALE_TIME.NONE,
       select: (queryData) =>
-        queryData.pages.flatMap((page) => page.posts).map(mapPostToCard),
+        queryData.pages
+          .flatMap((page) => page.posts)
+          .map((post) => mapPostToCard(post, COPY.htmlOnlyPost)),
     });
 
   const { targetRef } = useFetchNextOnScroll({
@@ -459,7 +355,10 @@ export const FeedPage = () => {
           <section className="tablet:pt-[120px] flex items-center justify-center px-4 pt-[160px] text-center">
             <div className="flex w-[176px] flex-col items-center gap-8">
               <p className="typo-body-16-regular whitespace-pre-line text-gray-500">
-                {formatEmptyResultMessage(trimmedDeferredKeyword)}
+                {formatEmptyResultMessage(
+                  trimmedDeferredKeyword,
+                  COPY.noSearchResult,
+                )}
               </p>
 
               <button
