+++ apps/web/src/shared/constants/html/index.ts
@@ -1,0 +1,1 @@
+export * from './pattern';
