+++ apps/web/src/shared/constants/html/pattern.ts
@@ -1,0 +1,1 @@
+export const HTML_TAG_PATTERN = /<\/?[a-z][\s\S]*?>/i;
